Rename api to matrix-api, using matrix-api crate in native master
Renamed api to matrix-api as it was foldered as api and in the
cargo.toml file messenger. I thought matrix api made more sense. feel
free to change.

Also add a main.rs to the native crate and imported in the matrix-api.
IT WORKS. So likely going to remove this in the next step so dw about
this file. Good to know we can do this though.