Add first draft of api client
We need to persist both the api base url as different servers will have
different url's and also a users access token to send as a bearer access
token.

To do this we are setting up a client struct that allows you to set and
remove a users access token etc. As I said is first draft so defs will
change! Compiles but not tested otherwise.