Pass through matrix client from main to all CI methods
Currently undecided on the best way to go forward however to share a
single instance of an object in rust seems to require passing it through
from main to the rest of the app. There is some singleton stuff you can
do and some things that will allow you to use it as a constant but
honestly not 100% sure which way is best to go so doing this for now
until we decide its causing issues.