Fix MatrixClient refs and allow to save access token
Previously the matrix client was not allowing me to save the updated
access token and was erroring on mutablity issues. I have fixed that
though not 100% sure as to why it works. Basically to you set the matrix
client to a mutable reference on creation and pass that down saving alot
of hassle allowing it to be changed.

Also added in something that showed the access token when you list
public rooms to make sure it was working (it was ;))